I need to fix the spdx-id.csv, because still there are double entries.
Theoretically I am starting to test the whole approach:

If LicenseName is in OSADL: continue LCV;
If not:
Perform StaticMapping,
If the StaticMapping provides with an SPDX-id, continue with LCV logic, 
if NOT perform DynamicMapping.

I am now trying to test the function IsASPDX, after having created a 
list with all the recognized SPDX-ids.